% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packet.R
\name{Bank_loan}
\alias{Bank_loan}
\title{A illustration dataset called Bank_loan}
\description{
A dataset in homework0.
}
\examples{
\dontrun{
data = read.table("../data/Bank_loan.txt",head=TRUE,na.strings = c("NA"))
past.customers = subset(data,default != "NA")
pie(c(sum(past.customers$default == 1),sum(past.customers$default == 0)),c(1,0))
}
}
