% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_for_project.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{rwMetropolisR}) and Cpp functions (\code{gibbsC} and \code{rwMetropolisC}).
}
\examples{
\dontrun{
data(data)
attach(data)
tm1 <- microbenchmark::microbenchmark(
  rnR = gibbsR(100,10),
  rnC = gibbsC(100,10)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
