% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_packet.R
\name{data}
\alias{data}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of \code{vaccR} and \code{vaccC}.
}
\examples{
\dontrun{
data(data)
attach(data)
tm <- microbenchmark::microbenchmark(
  vR = vaccR(age,female,ily),
  vC = vaccC(age,female,ily)
)
print(summary(tm)[,c(1,3,5,6)])
}
}
