% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{plot_s}
\alias{plot_s}
\title{A graph show about my simulation}
\usage{
plot_s(counter, result)
}
\arguments{
\item{counter}{indicator of which sigma it comes from}

\item{result}{the raw data about my simulation result}
}
\value{
a line chart including the tendecy of different methods I used
}
\description{
to instaintiate the table data by using the graph
}
\examples{
\dontrun{
result = array(data = 0, dim = c(10,4,9))
for (i in 1:10){
 a = read.table(paste("../data/text",i,".txt",sep = ""))
 a = a[,c(2,1,3,4,5,6,7,8,9)]
 result[i,,] = as.matrix(a)
 }
for (i in 1:4){
 par(mfrow = c(1,1))
 plot_s(i)
}
}
}
