% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{rtrunc}
\alias{rtrunc}
\title{random numbers for truncated distribution}
\usage{
rtrunc(n, distr, lower = -Inf, upper = Inf, ...)
}
\arguments{
\item{n}{the number of random numbers you want to generate}

\item{distr}{the distribution name}

\item{lower}{the left trunction}

\item{upper}{the right trunction}

\item{...}{some manual setting}
}
\value{
random numbers you want
}
\description{
to generate random numbers for truncated normal distribution in the function sim_result
}
\examples{
\dontrun{
x = rtrunc(10, "norm", lower = 0, mean = 2, sd = 5)
}
}
