% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{Use three inputs to predict response using Rcpp.}
\usage{
rwMetropolisC(n, sigma, x0, N, burn)
}
\arguments{
\item{n}{degree of freedom of t distribution}

\item{sigma}{standard variance of proposal distribution N(xt,sigma^2)}

\item{x0}{initial value}

\item{N}{size of random numbers required}

\item{burn}{burn-in length}
}
\value{
a random sample of size \code{N-burn}
}
\description{
The prediction model is described in http://www.babelgraph.org/wp/?p=358.
}
\examples{
\dontrun{
data(data)引用数据的方式
attach(data)
res <- test(n)
}
}
