% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{rwMetropolisR}
\alias{rwMetropolisR}
\title{A random walk Metropolis sampler using R for t distribution}
\usage{
rwMetropolisR(n, sigma, x0, N, burn)
}
\arguments{
\item{n}{degree of freedom of t distribution}

\item{sigma}{standard variance of proposal distribution N(xt,sigma^2)}

\item{x0}{initial value}

\item{N}{size of random numbers required}

\item{burn}{burn-in length}
}
\value{
a random sample of size \code{N-burn}
}
\description{
A random walk Metropolis sampler using R
}
\examples{
\dontrun{
X = rwMetropolisR(2,1,1,100,2) #n = 2
a = c(0.05,seq(0.1,0.9,0.1),0.95)
Q = qt(a,2)
qqplot(a,quantile(X,a))
}
}
