% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{sim_result}
\alias{sim_result}
\title{A function for the simulation in the paper}
\usage{
sim_result(n_replication, n, sigma2)
}
\arguments{
\item{n_replication}{number of replication}

\item{n}{number of data you already know}

\item{sigma2}{ture varaince in the simulation}
}
\value{
a line chart including the tendecy of different methods I used
}
\description{
to manipulate the 3 methods proposed in the paper and 6 nonparametric methods
}
\examples{
\dontrun{
begin = Sys.time()
n_total = c(5,10,20,100)
result = array(data = 0, dim = c(4,9))
n_replication = 1000
sigma2 = 1
for (i in 1:4){
     a = sim_result(n_replication,n,sigma2)
     result[i,] = as.matrix(a)
 }
end = Sys.time()
print(end-begin)
print(result)
}
}
