\name{StatComp22015-package}
\alias{StatComp22015-package}
\alias{StatComp22015}
\docType{package}
\title{
  a simple R package developed  for final project including several functions.
}
\description{
 Not only R and Rcpp versions of gibbs sampler and rwMetropolis sampler but also my simulations for the papers in the refferences are in the introduction. Homework rmd includes all my homework throgh the course.
}
\author{
22015
Maintainer: zyyfor2000<482225121@qq.com>
}
\references{
[1] PRAK D, TEUNTER R. A general method for addressing forecasting uncertainty in inventory models[J/OL]. International Journal of Forecasting, 2019, 35(1): 224-238. DOI:10.1016/j.ijforecast.2017.11.004.

[2] Sfakianakis M E, Verginis D G. A new family of nonparametric quantile estimators[J]. Communications in Statistics—Simulation and Computation®, 2008, 37(2): 337-345.
}
\keyword{ package }

