% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test}
\alias{test}
\title{Use three inputs to predict response using Rcpp.}
\usage{
test(n)
}
\arguments{
\item{n}{the first predictor (numeric)}
}
\value{
square of \code{n}
}
\description{
The prediction model is described in http://www.babelgraph.org/wp/?p=358.
}
\examples{
\dontrun{
data(data)引用数据的方式
attach(data)
res <- test(n)
}
}
